requirements.txt
# Core scientific computing libraries
numpy>=1.21.0
scipy>=1.7.0

# Computer vision library
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# Deep learning framework
torch>=1.9.0
torchvision>=0.10.0

# GUI framework
PySide6>=6.0.0

# Image processing library
Pillow>=8.3.0

# Symbolic mathematics library (used for switch function)
sympy>=1.8

# CSV file handling (built-in but listing for completeness)
csv>=1.0